backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.0
openai>=1.3.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
sentence-transformers>=2.2.0
psycopg2-binary>=2.9.0